+++ test/OpenTelemetry.Instrumentation.SqlClient.Tests/SqlClientIntegrationTests.cs
@@ -59,15 +59,15 @@ public sealed class SqlClientIntegrationTests : IClassFixture<SqlClientIntegrati
             })
             .Build();
 
-        using SqlConnection sqlConnection = new SqlConnection(this.GetConnectionString());
+        using var sqlConnection = new SqlConnection(this.GetConnectionString());
 
         sqlConnection.Open();
 
-        string dataSource = sqlConnection.DataSource;
+        var dataSource = sqlConnection.DataSource;
 
         sqlConnection.ChangeDatabase("master");
 #pragma warning disable CA2100
-        using SqlCommand sqlCommand = new SqlCommand(commandText, sqlConnection)
+        using var sqlCommand = new SqlCommand(commandText, sqlConnection)
 #pragma warning restore CA2100
         {
             CommandType = commandType,
@@ -90,13 +90,11 @@ public sealed class SqlClientIntegrationTests : IClassFixture<SqlClientIntegrati
         if (isFailure)
         {
 #if NET
-            var status = activity.GetStatus();
             Assert.Equal(ActivityStatusCode.Error, activity.Status);
             Assert.Equal("Divide by zero error encountered.", activity.StatusDescription);
             Assert.EndsWith("SqlException", activity.GetTagValue(SemanticConventions.AttributeErrorType) as string);
             Assert.Equal("8134", activity.GetTagValue(SemanticConventions.AttributeDbResponseStatusCode));
 #else
-            var status = activity.GetStatus();
             Assert.Equal(ActivityStatusCode.Error, activity.Status);
             Assert.Equal("8134", activity.StatusDescription);
             Assert.EndsWith("SqlException", activity.GetTagValue(SemanticConventions.AttributeErrorType) as string);
@@ -107,14 +105,11 @@ public sealed class SqlClientIntegrationTests : IClassFixture<SqlClientIntegrati
 
     private string GetConnectionString()
     {
-        switch (this.fixture.DatabaseContainer)
+        return this.fixture.DatabaseContainer switch
         {
-            case SqlEdgeContainer container:
-                return container.GetConnectionString();
-            case MsSqlContainer container:
-                return container.GetConnectionString();
-            default:
-                throw new InvalidOperationException($"Container type '${this.fixture.DatabaseContainer.GetType().Name}' is not supported.");
-        }
+            SqlEdgeContainer container => container.GetConnectionString(),
+            MsSqlContainer container => container.GetConnectionString(),
+            _ => throw new InvalidOperationException($"Container type '${this.fixture.DatabaseContainer.GetType().Name}' is not supported."),
+        };
     }
 }