+++ test/OpenTelemetry.Instrumentation.SqlClient.Tests/SqlClientTests.cs
@@ -6,6 +6,8 @@ using System.Diagnostics;
 using Microsoft.Data.SqlClient;
 using Microsoft.Extensions.DependencyInjection;
 using OpenTelemetry.Instrumentation.SqlClient.Implementation;
+using OpenTelemetry.Metrics;
+
 #if !NETFRAMEWORK
 using OpenTelemetry.Tests;
 #endif
@@ -44,8 +46,8 @@ public class SqlClientTests : IDisposable
     [Fact]
     public void SqlClient_NamedOptions()
     {
-        int defaultExporterOptionsConfigureOptionsInvocations = 0;
-        int namedExporterOptionsConfigureOptionsInvocations = 0;
+        var defaultExporterOptionsConfigureOptionsInvocations = 0;
+        var namedExporterOptionsConfigureOptionsInvocations = 0;
 
         using var tracerProvider = Sdk.CreateTracerProviderBuilder()
             .ConfigureServices(services =>
@@ -65,34 +67,7 @@ public class SqlClientTests : IDisposable
     // DiagnosticListener-based instrumentation is only available on .NET Core
 #if !NETFRAMEWORK
     [Theory]
-    [InlineData(SqlClientDiagnosticListener.SqlDataBeforeExecuteCommand, SqlClientDiagnosticListener.SqlDataAfterExecuteCommand, CommandType.StoredProcedure, "SP_GetOrders", false)]
-    [InlineData(SqlClientDiagnosticListener.SqlDataBeforeExecuteCommand, SqlClientDiagnosticListener.SqlDataAfterExecuteCommand, CommandType.StoredProcedure, "SP_GetOrders", false, false)]
-    [InlineData(SqlClientDiagnosticListener.SqlDataBeforeExecuteCommand, SqlClientDiagnosticListener.SqlDataAfterExecuteCommand, CommandType.Text, "select * from sys.databases", false)]
-    [InlineData(SqlClientDiagnosticListener.SqlDataBeforeExecuteCommand, SqlClientDiagnosticListener.SqlDataAfterExecuteCommand, CommandType.Text, "select * from sys.databases", false, false)]
-    [InlineData(SqlClientDiagnosticListener.SqlMicrosoftBeforeExecuteCommand, SqlClientDiagnosticListener.SqlMicrosoftAfterExecuteCommand, CommandType.StoredProcedure, "SP_GetOrders", true)]
-    [InlineData(SqlClientDiagnosticListener.SqlMicrosoftBeforeExecuteCommand, SqlClientDiagnosticListener.SqlMicrosoftAfterExecuteCommand, CommandType.StoredProcedure, "SP_GetOrders", true, false)]
-    [InlineData(SqlClientDiagnosticListener.SqlMicrosoftBeforeExecuteCommand, SqlClientDiagnosticListener.SqlMicrosoftAfterExecuteCommand, CommandType.Text, "select * from sys.databases", true)]
-    [InlineData(SqlClientDiagnosticListener.SqlMicrosoftBeforeExecuteCommand, SqlClientDiagnosticListener.SqlMicrosoftAfterExecuteCommand, CommandType.Text, "select * from sys.databases", true, false)]
-
-    // Test cases when EmitOldAttributes = false and EmitNewAttributes = true (i.e., OTEL_SEMCONV_STABILITY_OPT_IN=database)
-    [InlineData(SqlClientDiagnosticListener.SqlDataBeforeExecuteCommand, SqlClientDiagnosticListener.SqlDataAfterExecuteCommand, CommandType.StoredProcedure, "SP_GetOrders", false, true, false, true)]
-    [InlineData(SqlClientDiagnosticListener.SqlDataBeforeExecuteCommand, SqlClientDiagnosticListener.SqlDataAfterExecuteCommand, CommandType.StoredProcedure, "SP_GetOrders", false, false, false, true)]
-    [InlineData(SqlClientDiagnosticListener.SqlDataBeforeExecuteCommand, SqlClientDiagnosticListener.SqlDataAfterExecuteCommand, CommandType.Text, "select * from sys.databases", false, true, false, true)]
-    [InlineData(SqlClientDiagnosticListener.SqlDataBeforeExecuteCommand, SqlClientDiagnosticListener.SqlDataAfterExecuteCommand, CommandType.Text, "select * from sys.databases", false, false, false, true)]
-    [InlineData(SqlClientDiagnosticListener.SqlMicrosoftBeforeExecuteCommand, SqlClientDiagnosticListener.SqlMicrosoftAfterExecuteCommand, CommandType.StoredProcedure, "SP_GetOrders", true, true, false, true)]
-    [InlineData(SqlClientDiagnosticListener.SqlMicrosoftBeforeExecuteCommand, SqlClientDiagnosticListener.SqlMicrosoftAfterExecuteCommand, CommandType.StoredProcedure, "SP_GetOrders", true, false, false, true)]
-    [InlineData(SqlClientDiagnosticListener.SqlMicrosoftBeforeExecuteCommand, SqlClientDiagnosticListener.SqlMicrosoftAfterExecuteCommand, CommandType.Text, "select * from sys.databases", true, true, false, true)]
-    [InlineData(SqlClientDiagnosticListener.SqlMicrosoftBeforeExecuteCommand, SqlClientDiagnosticListener.SqlMicrosoftAfterExecuteCommand, CommandType.Text, "select * from sys.databases", true, false, false, true)]
-
-    // Test cases when EmitOldAttributes = true and EmitNewAttributes = true (i.e., OTEL_SEMCONV_STABILITY_OPT_IN=database/dup)
-    [InlineData(SqlClientDiagnosticListener.SqlDataBeforeExecuteCommand, SqlClientDiagnosticListener.SqlDataAfterExecuteCommand, CommandType.StoredProcedure, "SP_GetOrders", false, true, true, true)]
-    [InlineData(SqlClientDiagnosticListener.SqlDataBeforeExecuteCommand, SqlClientDiagnosticListener.SqlDataAfterExecuteCommand, CommandType.StoredProcedure, "SP_GetOrders", false, false, true, true)]
-    [InlineData(SqlClientDiagnosticListener.SqlDataBeforeExecuteCommand, SqlClientDiagnosticListener.SqlDataAfterExecuteCommand, CommandType.Text, "select * from sys.databases", false, true, true, true)]
-    [InlineData(SqlClientDiagnosticListener.SqlDataBeforeExecuteCommand, SqlClientDiagnosticListener.SqlDataAfterExecuteCommand, CommandType.Text, "select * from sys.databases", false, false, true, true)]
-    [InlineData(SqlClientDiagnosticListener.SqlMicrosoftBeforeExecuteCommand, SqlClientDiagnosticListener.SqlMicrosoftAfterExecuteCommand, CommandType.StoredProcedure, "SP_GetOrders", true, true, true, true)]
-    [InlineData(SqlClientDiagnosticListener.SqlMicrosoftBeforeExecuteCommand, SqlClientDiagnosticListener.SqlMicrosoftAfterExecuteCommand, CommandType.StoredProcedure, "SP_GetOrders", true, false, true, true)]
-    [InlineData(SqlClientDiagnosticListener.SqlMicrosoftBeforeExecuteCommand, SqlClientDiagnosticListener.SqlMicrosoftAfterExecuteCommand, CommandType.Text, "select * from sys.databases", true, true, true, true)]
-    [InlineData(SqlClientDiagnosticListener.SqlMicrosoftBeforeExecuteCommand, SqlClientDiagnosticListener.SqlMicrosoftAfterExecuteCommand, CommandType.Text, "select * from sys.databases", true, false, true, true)]
+    [MemberData(nameof(SqlTestData.SqlClientCallsAreCollectedSuccessfullyCases), MemberType = typeof(SqlTestData))]
     public void SqlClientCallsAreCollectedSuccessfully(
         string beforeCommand,
         string afterCommand,
@@ -101,27 +76,47 @@ public class SqlClientTests : IDisposable
         bool captureTextCommandContent,
         bool shouldEnrich = true,
         bool emitOldAttributes = true,
-        bool emitNewAttributes = false)
+        bool emitNewAttributes = false,
+        bool tracingEnabled = true,
+        bool metricsEnabled = true)
     {
         using var sqlConnection = new SqlConnection(TestConnectionString);
         using var sqlCommand = sqlConnection.CreateCommand();
 
         var activities = new List<Activity>();
-        using (Sdk.CreateTracerProviderBuilder()
-                .AddSqlClientInstrumentation(
-                    (opt) =>
-                    {
-                        opt.SetDbStatementForText = captureTextCommandContent;
-                        if (shouldEnrich)
-                        {
-                            opt.Enrich = ActivityEnrichment;
-                        }
+        var metrics = new List<Metric>();
 
-                        opt.EmitOldAttributes = emitOldAttributes;
-                        opt.EmitNewAttributes = emitNewAttributes;
-                    })
-                .AddInMemoryExporter(activities)
-                .Build())
+        var traceProviderBuilder = Sdk.CreateTracerProviderBuilder();
+
+        if (tracingEnabled)
+        {
+            traceProviderBuilder.AddSqlClientInstrumentation(
+            (opt) =>
+            {
+                opt.SetDbStatementForText = captureTextCommandContent;
+                if (shouldEnrich)
+                {
+                    opt.Enrich = ActivityEnrichment;
+                }
+
+                opt.EmitOldAttributes = emitOldAttributes;
+                opt.EmitNewAttributes = emitNewAttributes;
+            });
+            traceProviderBuilder.AddInMemoryExporter(activities);
+        }
+
+        var meterProviderBuilder = Sdk.CreateMeterProviderBuilder();
+
+        if (metricsEnabled)
+        {
+            meterProviderBuilder.AddSqlClientInstrumentation();
+            meterProviderBuilder.AddInMemoryExporter(metrics);
+        }
+
+        var traceProvider = traceProviderBuilder.Build();
+        var meterProvider = meterProviderBuilder.Build();
+
+        try
         {
             var operationId = Guid.NewGuid();
             sqlCommand.CommandType = commandType;
@@ -151,20 +146,42 @@ public class SqlClientTests : IDisposable
                 afterCommand,
                 afterExecuteEventData);
         }
+        finally
+        {
+            traceProvider.Dispose();
+            meterProvider.Dispose();
+        }
 
-        Assert.Single(activities);
-        var activity = activities[0];
+        Activity? activity = null;
 
-        VerifyActivityData(
-            sqlCommand.CommandType,
-            sqlCommand.CommandText,
-            captureTextCommandContent,
-            false,
-            false,
-            shouldEnrich,
-            activity,
-            emitOldAttributes,
-            emitNewAttributes);
+        if (tracingEnabled)
+        {
+            activity = Assert.Single(activities);
+            VerifyActivityData(
+                sqlCommand.CommandType,
+                sqlCommand.CommandText,
+                captureTextCommandContent,
+                false,
+                false,
+                shouldEnrich,
+                activity,
+                emitOldAttributes,
+                emitNewAttributes);
+        }
+
+        var dbClientOperationDurationMetrics = metrics
+            .Where(metric => metric.Name == "db.client.operation.duration")
+            .ToArray();
+
+        if (metricsEnabled)
+        {
+            var metric = Assert.Single(dbClientOperationDurationMetrics);
+            VerifyDurationMetricData(metric, activity);
+        }
+        else
+        {
+            Assert.Empty(dbClientOperationDurationMetrics);
+        }
     }
 
     [Theory]
@@ -221,20 +238,25 @@ public class SqlClientTests : IDisposable
     }
 
     [Theory]
-    [InlineData(SqlClientDiagnosticListener.SqlDataBeforeExecuteCommand, SqlClientDiagnosticListener.SqlDataWriteCommandError)]
-    [InlineData(SqlClientDiagnosticListener.SqlDataBeforeExecuteCommand, SqlClientDiagnosticListener.SqlDataWriteCommandError, false)]
-    [InlineData(SqlClientDiagnosticListener.SqlDataBeforeExecuteCommand, SqlClientDiagnosticListener.SqlDataWriteCommandError, false, true)]
-    [InlineData(SqlClientDiagnosticListener.SqlMicrosoftBeforeExecuteCommand, SqlClientDiagnosticListener.SqlMicrosoftWriteCommandError)]
-    [InlineData(SqlClientDiagnosticListener.SqlMicrosoftBeforeExecuteCommand, SqlClientDiagnosticListener.SqlMicrosoftWriteCommandError, false)]
-    [InlineData(SqlClientDiagnosticListener.SqlMicrosoftBeforeExecuteCommand, SqlClientDiagnosticListener.SqlMicrosoftWriteCommandError, false, true)]
-    public void SqlClientErrorsAreCollectedSuccessfully(string beforeCommand, string errorCommand, bool shouldEnrich = true, bool recordException = false)
+    [MemberData(nameof(SqlTestData.SqlClientErrorsAreCollectedSuccessfullyCases), MemberType = typeof(SqlTestData))]
+    public void SqlClientErrorsAreCollectedSuccessfully(
+        string beforeCommand,
+        string errorCommand,
+        bool shouldEnrich = true,
+        bool recordException = false,
+        bool tracingEnabled = true,
+        bool metricsEnabled = true)
     {
         using var sqlConnection = new SqlConnection(TestConnectionString);
         using var sqlCommand = sqlConnection.CreateCommand();
 
         var activities = new List<Activity>();
-        using (Sdk.CreateTracerProviderBuilder()
-            .AddSqlClientInstrumentation(options =>
+        var metrics = new List<Metric>();
+        var traceProviderBuilder = Sdk.CreateTracerProviderBuilder();
+
+        if (tracingEnabled)
+        {
+            traceProviderBuilder.AddSqlClientInstrumentation(options =>
             {
                 options.RecordException = recordException;
                 if (shouldEnrich)
@@ -242,8 +264,23 @@ public class SqlClientTests : IDisposable
                     options.Enrich = ActivityEnrichment;
                 }
             })
-            .AddInMemoryExporter(activities)
-            .Build())
+            .AddInMemoryExporter(activities);
+        }
+
+        var traceProvider = traceProviderBuilder.Build();
+
+        var meterProviderBuilder = Sdk.CreateMeterProviderBuilder();
+
+        if (metricsEnabled)
+        {
+            meterProviderBuilder
+                .AddSqlClientInstrumentation()
+                .AddInMemoryExporter(metrics);
+        }
+
+        var meterProvider = meterProviderBuilder.Build();
+
+        try
         {
             var operationId = Guid.NewGuid();
             sqlCommand.CommandText = "SP_GetOrders";
@@ -272,18 +309,44 @@ public class SqlClientTests : IDisposable
                 errorCommand,
                 commandErrorEventData);
         }
+        finally
+        {
+            traceProvider.Dispose();
+            meterProvider.Dispose();
+        }
 
-        Assert.Single(activities);
-        var activity = activities[0];
+        Activity? activity = null;
 
-        VerifyActivityData(
-            sqlCommand.CommandType,
-            sqlCommand.CommandText,
-            false,
-            true,
-            recordException,
-            shouldEnrich,
-            activity);
+        if (tracingEnabled)
+        {
+            activity = Assert.Single(activities);
+            VerifyActivityData(
+                sqlCommand.CommandType,
+                sqlCommand.CommandText,
+                false,
+                true,
+                recordException,
+                shouldEnrich,
+                activity);
+        }
+        else
+        {
+            Assert.Empty(activities);
+        }
+
+        var dbClientOperationDurationMetrics = metrics
+            .Where(metric => metric.Name == "db.client.operation.duration")
+            .ToArray();
+
+        if (metricsEnabled)
+        {
+            var metric = Assert.Single(dbClientOperationDurationMetrics);
+            VerifyDurationMetricData(metric, activity);
+        }
+        else
+        {
+            Assert.Empty(dbClientOperationDurationMetrics);
+        }
     }
 
     [Theory]
@@ -310,12 +373,7 @@ public class SqlClientTests : IDisposable
             },
             cmd =>
             {
-                if (cmd is SqlCommand command)
-                {
-                    return command.CommandText == "select 2";
-                }
-
-                return true;
+                return cmd is not SqlCommand command || command.CommandText == "select 2";
             });
 
         Assert.Single(activities);
@@ -333,12 +391,7 @@ public class SqlClientTests : IDisposable
             },
             cmd =>
             {
-                if (cmd is SqlCommand command)
-                {
-                    return command.CommandText == "select 2";
-                }
-
-                return true;
+                return cmd is not SqlCommand command || command.CommandText == "select 2";
             });
 
         Assert.Empty(activities);
@@ -386,7 +439,6 @@ public class SqlClientTests : IDisposable
         }
         else
         {
-            var status = activity.GetStatus();
             Assert.Equal(ActivityStatusCode.Error, activity.Status);
             Assert.NotNull(activity.StatusDescription);
 
@@ -406,7 +458,7 @@ public class SqlClientTests : IDisposable
         if (shouldEnrich)
         {
             Assert.NotEmpty(activity.Tags.Where(tag => tag.Key == "enriched"));
-            Assert.Equal("yes", activity.Tags.Where(tag => tag.Key == "enriched").FirstOrDefault().Value);
+            Assert.Equal("yes", activity.Tags.FirstOrDefault(tag => tag.Key == "enriched").Value);
         }
         else
         {
@@ -462,6 +514,34 @@ public class SqlClientTests : IDisposable
                 }
 
                 break;
+            case CommandType.TableDirect:
+                Assert.Fail("Not supported command type: CommandType.TableDirect");
+                break;
+            default:
+                Assert.Fail($"Not supported command type: {commandType}");
+                break;
+        }
+    }
+
+    internal static void VerifyDurationMetricData(Metric metric, Activity? activity)
+    {
+        Assert.NotNull(metric);
+        Assert.Equal("s", metric.Unit);
+        Assert.Equal(MetricType.Histogram, metric.MetricType);
+
+        var metricPoints = new List<MetricPoint>();
+        foreach (var p in metric.GetMetricPoints())
+        {
+            metricPoints.Add(p);
+        }
+
+        var metricPoint = Assert.Single(metricPoints);
+
+        if (activity != null)
+        {
+            var count = metricPoint.GetHistogramCount();
+            var sum = metricPoint.GetHistogramSum();
+            Assert.Equal(activity.Duration.TotalSeconds, sum);
         }
     }
 
@@ -614,7 +694,7 @@ public class SqlClientTests : IDisposable
                 afterExecuteEventData);
         }
 
-        return activities.ToArray();
+        return [.. activities];
     }
 #endif