+++ src/OpenTelemetry.Exporter.Geneva/CHANGELOG.md
@@ -19,6 +19,9 @@
   For configuration details see:
   [OtlpProtobufEncoding](./README.md#otlpprotobufencoding).
 
+* Update OpenTelemetry SDK version to `1.10.0`.
+  ([#2317](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2317))
+
 ## 1.9.0
 
 Released 2024-Jun-21