+++ src/OpenTelemetry.Resources.Container/CHANGELOG.md
@@ -6,6 +6,9 @@
   and add .NET 8/.NET Standard 2.0 targets.
   ([#2166](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2166))
 
+* Updated OpenTelemetry core component version(s) to `1.10.0`.
+  ([#2317](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2317))
+
 * Add Kubernetes support.
   ([#1699](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1699))